Split up the ValidationService into several parts and its own package.
