+++ src/main/java/uk/nhs/ctp/tkwvalidation/AuditSelector.java
@@ -1,0 +1,57 @@
+package uk.nhs.ctp.tkwvalidation;
+
+import static org.springframework.web.bind.annotation.RequestMethod.GET;
+import static org.springframework.web.bind.annotation.RequestMethod.POST;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
+import java.util.stream.Stream;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.RequestMethod;
+import uk.nhs.ctp.audit.model.AuditSession;
+import uk.nhs.ctp.auditFinder.model.OperationType;
+import uk.nhs.ctp.tkwvalidation.models.HttpMessageAudit;
+
+@Component
+public class AuditSelector {
+
+  private static final String CASE_ID = "caseId";
+
+  public List<HttpMessageAudit> selectAudits(
+      Collection<AuditSession> audits,
+      OperationType operationType) {
+    var zippableAudits = new ArrayList<HttpMessageAudit>();
+
+    for (var audit : audits) {
+      var baseName = operationType == OperationType.SERVICE_SEARCH
+          ? operationType.getName()
+          : "encounter" + audit.getAdditionalProperties().get(CASE_ID);
+
+      if (isMethod(audit.getRequestMethod(), POST)) {
+        zippableAudits.add(HttpMessageAudit.fromRequest(audit, baseName));
+      }
+      if (isMethod(audit.getRequestMethod(), GET, POST)) {
+        zippableAudits.add(HttpMessageAudit.fromResponse(audit, baseName));
+      }
+
+      for (var entry : audit.getEntries()) {
+        if (isMethod(entry.getRequestMethod(), POST)) {
+          zippableAudits.add(HttpMessageAudit.fromRequest(entry, baseName));
+        }
+        if (isMethod(entry.getRequestMethod(), GET, POST)) {
+          zippableAudits.add(HttpMessageAudit.fromResponse(entry, baseName));
+        }
+      }
+    }
+
+    zippableAudits.sort(Comparator.comparing(HttpMessageAudit::getMoment));
+    return zippableAudits;
+  }
+
+  private static boolean isMethod(String method, RequestMethod... expectedMethods) {
+    return Stream.of(expectedMethods).map(Enum::name).anyMatch(method::equalsIgnoreCase);
+  }
+
+}
