+++ src/main/java/uk/nhs/ctp/tkwvalidation/ZipBuilder.java
@@ -1,4 +1,4 @@
-package uk.nhs.ctp.utils;
+package uk.nhs.ctp.tkwvalidation;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
