+++ src/main/java/uk/nhs/ctp/tkwvalidation/models/HttpMessageType.java
@@ -1,0 +1,6 @@
+package uk.nhs.ctp.tkwvalidation.models;
+
+public enum HttpMessageType {
+  REQUEST,
+  RESPONSE
+}
