+++ src/main/java/uk/nhs/ctp/tkwvalidation/rules/AuditValidationRule.java
@@ -1,0 +1,8 @@
+package uk.nhs.ctp.tkwvalidation.rules;
+
+import java.util.List;
+import uk.nhs.ctp.audit.model.AuditSession;
+
+public interface AuditValidationRule {
+  void ensure(List<AuditSession> audits);
+}
